requirements.txt
pandas
numpy
matplotlib
scikit-learn
notebook
python-box==6.0.2
pyYAML
tqdm
scipy
Flask-Cors
seaborn
Flask
dvc
tensorflow
pylint
black 
pytest
ensure
joblib


# -e .
